requirements.txt
asgiref==3.4.1
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
